[add] vite-plugin-pwaを使ってservice worker のみ使う設定を追加
(cherry picked from commit c321ee26a440996de18be7925b1acd5eddb7eb5c)
